refactor : useProducts 생성
